FS-5062 - Transferring tests of generate_fund_round_config module from general test_config_export into separate test module
